+++ src/App.tsx
@@ -1,5 +1,5 @@
 import PokemonDetailsApp from 'features/pokemon-details/app';
-import PokemonTableApp from 'features/pokemon-table/app';
+import PokemonApp from 'features/pokemon/app';
 import { BrowserRouter, Route, Routes } from 'react-router';
 import './App.css';
 
@@ -7,7 +7,7 @@ const App = (): React.ReactNode => {
   return (
     <BrowserRouter>
       <Routes>
-        <Route path='/' element={<PokemonTableApp />} />
+        <Route path='/' element={<PokemonApp />} />
         <Route path='/pokemon-details/:name' element={<PokemonDetailsApp />} />
       </Routes>
     </BrowserRouter>
